allow no-admix
